% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_setting}
\alias{create_setting}
\title{Define a simulation setting where some nuisance quantities are consistently estimated}
\usage{
create_setting(which_correct)
}
\arguments{
\item{which_correct}{A character vector indicating which nuisance regressions are consistently estimated.}
}
\value{
A named list that can be passed to \code{one_sim} function to run a simulation
under a certain pattern of nuisance parameter (mis)specification.
}
\description{
If a nuisance quantity is included in \code{which_correct}, the proper specification is
given for the regression function. If a nuisance quantity is not included in \code{which_correct},
an intercept-only model is used instead.
}
