% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_NA}
\alias{add_NA}
\title{Add NA value back indicating seed position}
\usage{
add_NA(vec, seed_position)
}
\arguments{
\item{vec}{A vector indicating the MoCo estimates, adjusted association, z-scores, or significant regions.}

\item{seed_position}{Position indicator of where the seed region is.}
}
\value{
A vector with NA values inserted to indicate the position of the seed region.
}
\description{
Add NA value back indicating seed position
}
