% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypo_test.R
\name{hypo_test}
\alias{hypo_test}
\title{Simultaneous Confidence Interval Hypothesis Testing}
\usage{
hypo_test(result, fwer = 0.05, seed = 1)
}
\arguments{
\item{result}{The result of fitting based on the one-step function or one-step cross function.}

\item{fwer}{Family-wise error rates (FWER) for control.}

\item{seed}{A numeric value to control the randomness of the random variable sampling.}
}
\value{
A list with named entries:
\item{conf_band}{Simultaneous confidence band based on specified FWER.}
\item{z_score}{A p-length vector of z-scores for each functional connectivity.}
\item{significant_regions}{A p-length vector of TRUE or FALSE, indicating significance for each functional connectivity.}
}
\description{
Simultaneous Confidence Interval Hypothesis Testing
}
\details{
Conducts hypothesis testing for significant associations using simultaneous confidence intervals.
}
